Add files via upload
This challenge felt a lot easier, my biggest learning curve was how to iterate through the list and give absolute locations without re-indexing the same thing over and over or removing elements to cause a bad count. The solution I found was using a for loop and taking the variable to splice the list.